+++ app/build.gradle.kts
@@ -33,7 +33,6 @@ android {
     }
 
     compileOptions {
-        isCoreLibraryDesugaringEnabled = true
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
     }
@@ -56,7 +55,6 @@ dependencies {
     implementation("androidx.compose.material3:material3:1.0.0-alpha15")
     implementation("androidx.navigation:navigation-compose:2.5.1")
     // jetpack compose
-    coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:1.1.5") // <- this dependency is required
     implementation(Deps.Compose.ui)
     implementation(Deps.Compose.material)
     implementation(Deps.Compose.activity)
